# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News API",
        "description": "API для керування новинами",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8082/api",
            "description": "Local server"
        }
    ],
    "paths": {
        "/admin/users": {
            "get": {
                "summary": "Список користувачів",
                "operationId": "2a04f4bb4444a2a498cd05658e60cffa",
                "responses": {
                    "200": {
                        "description": "Список користувачів"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}/role": {
            "post": {
                "summary": "Змінити роль користувача",
                "operationId": "251329f59d80da7abe1298a1965609a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "role"
                                ],
                                "properties": {
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "helper",
                                            "reader"
                                        ],
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Роль оновлено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/settings": {
            "get": {
                "summary": "Отримати налаштування",
                "operationId": "654187ac3fdae3a313708749481c1676",
                "responses": {
                    "200": {
                        "description": "Налаштування"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Оновити налаштування",
                "operationId": "195edd9635d772a8e3b0b39338d909e7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "site_name": {
                                        "type": "string",
                                        "example": "News Portal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Налаштування оновлено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/contacts": {
            "get": {
                "summary": "Список повідомлень",
                "operationId": "d2483bb379be3c4e2a9b38ea18ea3db9",
                "responses": {
                    "200": {
                        "description": "Список повідомлень"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/contacts/{id}": {
            "get": {
                "summary": "Деталі повідомлення",
                "operationId": "75a2da09ed87e9a17c8ed4c38417924f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Деталі повідомлення"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/contacts/{id}/status": {
            "patch": {
                "summary": "Оновити статус повідомлення",
                "operationId": "bdb937eb935044210d59abcafd88dd24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "new",
                                            "read",
                                            "replied"
                                        ],
                                        "example": "read"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Статус оновлено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/contacts/{id}/mark-read": {
            "patch": {
                "summary": "Позначити прочитаним",
                "operationId": "d649710277f7739346a730957680b782",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Позначено прочитаним"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/stats": {
            "get": {
                "summary": "Статистика",
                "operationId": "aea495bb05f23b43c987e8b21380cc66",
                "responses": {
                    "200": {
                        "description": "Статистика сайту"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}